requirements.txt
opencv-python
numpy
pyserial
customtkinter==5.2.2
CTkSpinbox==1.4.1
pillow==10.3.0
tkinter